libs: make every implicit integer conversion explicit
Preparation for promoting -Wconversion to a hard error (#240). Five
mechanical classes, all semantics-preserving:

- register accessors added ra8_*_addr_t + ra8_*_off_t enums directly;
  cast both operands to uintptr_t (clang -Wenum-enum-conversion), same
  for the other mixed-enum arithmetic in ra8_rar / ra8_rsip / ra8_usb_hhid
- 1UL/0x3UL shift-mask idioms stored into uint32_t are 64-bit on the
  unit-test host (clang/gcc -Wshorten-64-to-32 / -Wconversion); use the
  32-bit 1U/0x3U literals the target semantics always assumed
- 0U/1U ternaries assigned to uint8_t promote to int; cast the result
- ra8_reflow_internal_xhtml_invalid took uint32_t while its only caller
  passes the size_t engine->xhtml_len; widen the parameter to size_t
  (and include stddef.h in the internal header)
- complete the ra8_fs_backend_t initializer in exfat_fs_test (erase_blocks
  hook was silently dropped by the 4-element list)

+++ libs/ra8_comic/src/ra8_rar.c
@@ -302,7 +302,8 @@ static bool s_rar4_file(const ra8_rar_t* rar,
   size_t         pos    = (size_t)k_rar4_off_name;
   if (large) {
     pack |= (uint64_t)s_le32(&scratch[k_rar4_off_name]) << (uint32_t)k_rar_hi_sh;
-    unp |= (uint64_t)s_le32(&scratch[k_rar4_off_name + k_rar_w32]) << (uint32_t)k_rar_hi_sh;
+    unp |= (uint64_t)s_le32(&scratch[(size_t)k_rar4_off_name + (size_t)k_rar_w32])
+           << (uint32_t)k_rar_hi_sh;
     pos = (size_t)k_rar4_off_name_large;
   }
   out->is_file = 1U;
@@ -356,7 +357,7 @@ s_rar4_block(const ra8_rar_t* rar, uint64_t off, char* nbuf, uint16_t ncap, ra8_
   const bool islong = (flags & (uint16_t)k_rar4_flag_long) != 0U;
   uint64_t   add    = 0U;
   if (islong) {
-    if (n >= (size_t)(k_rar4_off_add + k_rar_w32)) {
+    if (n >= ((size_t)k_rar4_off_add + (size_t)k_rar_w32)) {
       add = (uint64_t)s_le32(&scratch[k_rar4_off_add]);
     }
   }
